refactor: extract styles and reorganize file structure
