+++ index.html
@@ -7,7 +7,7 @@
   </head>
   <body>
     <div id="app">
-      <my-app></my-app>
+      <my-menu></my-menu>
       <div id="outlet"></div>
     </div>
     <script type="module" src="/src/main.ts"></script>
