+++ src/components/bots/poet/poet-bot.ts
@@ -1,8 +1,11 @@
-import { css, html } from 'lit';
+import { html } from 'lit';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
+import { when } from 'lit-html/directives/when.js';
 import { customElement, property } from 'lit/decorators.js';
-import { router } from '../router';
-import { Bot } from './bot';
+
+import { router } from '../../../router';
+import { Bot } from '../bot';
+import { styles } from './poet-bot.styles';
 
 /**
  * Represents a Poet Bot that generates IT related poems.
@@ -17,75 +20,7 @@ export class PoetBot extends Bot {
   /**
    * The CSS styles for the Poet Bot component.
    */
-  static styles = css`
-    :host {
-      display: block;
-      font-family: 'Arial', sans-serif;
-      max-width: 600px;
-      margin: 2rem auto;
-      padding: 1rem 1.5rem;
-      background: #ffffff;
-      border-radius: 12px;
-      box-shadow: 0 6px 12px rgba(0, 0, 0, 0.1);
-      border: 1px solid #eaeaea;
-    }
-
-    h2 {
-      font-size: 1.25rem;
-      color: #222;
-      text-align: center;
-      margin-bottom: 1.5rem;
-    }
-
-    #botContainer {
-      display: flex;
-      flex-direction: column;
-      gap: 1.5rem;
-    }
-
-    #botQuestion {
-      font-size: 0.875rem;
-      color: #555;
-      margin-top: 2rem;
-      text-align: center;
-    }
-
-    #botResponse {
-      margin-top: 1rem;
-      padding: 1rem;
-      background: #f9f9f9;
-      border: 1px solid #ddd;
-      border-radius: 8px;
-      color: #333;
-      white-space: pre-wrap;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    }
-
-    button {
-      padding: 0.75rem 1.5rem;
-      font-size: 1rem;
-      color: #fff;
-      background-color: #007bff;
-      border: none;
-      border-radius: 8px;
-      cursor: pointer;
-      transition:
-        background-color 0.3s ease,
-        transform 0.2s ease;
-      align-self: center;
-    }
-
-    button:hover {
-      background-color: #0056b3;
-      transform: translateY(-2px);
-    }
-
-    .invisible {
-      opacity: 0;
-    }
-  `;
+  static styles = styles;
 
   constructor() {
     super('Poet bot');
@@ -141,9 +76,11 @@ export class PoetBot extends Bot {
       <div id="botContainer">
         <button @click="${this.start}">Generate Poem</button>
         <p id="botQuestion">${this.userQuestion.content || 'Subject will be generated...'}</p>
-        <div ?hidden=${this.response.length === 0} id="botResponse">
-          ${unsafeHTML(this.response)}
-        </div>
+
+        ${when(
+          this.response.length > 0,
+          () => html`<div id="botResponse">${unsafeHTML(this.response)}</div>`
+        )}
       </div>
     `;
   }
