+++ src/components/home.ts
@@ -1,8 +1,9 @@
 // login-component.ts
 import { RouterLocation } from '@vaadin/router';
-import { LitElement, css, html } from 'lit';
+import { LitElement, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
-import { router } from './router';
+
+import { router } from '../router';
 
 @customElement('home-component')
 export class HomeComponent extends LitElement {
@@ -13,11 +14,7 @@ export class HomeComponent extends LitElement {
     this.location = router.location;
   }
 
-  static styles = css`
-    /* Add your styles here */
-  `;
-
   render() {
-    return html` <slot></slot> `;
+    return html`<slot></slot>`;
   }
 }
